Fixed QA issues

+++ includes/ajax/investment/ticker.php
@@ -30,13 +30,13 @@ try {
             if(!empty($_POST['type'])&&$_POST['type']=="coin"){
                 $token_price['data']['amount'] = round($_POST['total_tokens']*$token_price['data']['buy_price'], 2);
                 $token_price['data']['sub_total'] = round($_POST['total_tokens']*$token_price['data']['buy_price'], 2);
-                $token_price['data']['total_fees'] = round($_POST['total_tokens']*$fees/100, 5).' '.$_POST['token'];
+                $token_price['data']['total_fees'] = (float)round($_POST['total_tokens']*$fees/100, 5).' '.$_POST['token'];
                 // $token_price['data']['selling_sub_total'] = number_format((double)$_POST['token']*(double)$token_price['data']['selling_fees'], 5, '.', '');
                 // echo '<pre>'; print_r($token_price); die;
             }else{
                 $token_price['data']['tokens'] = round($_POST['amount']/$token_price['data']['buy_price'], 5);
                 $token_price['data']['sub_total'] = $_POST['amount'];
-                $token_price['data']['total_fees'] = round($token_price['data']['tokens']*$fees/100, 5).' '.$_POST['token'];
+                $token_price['data']['total_fees'] = (float)round($token_price['data']['tokens']*$fees/100, 5).' '.$_POST['token'];
             }
             break;
         case 'sell':
